put custom admin setting components into their own files.
improve code comments and class names while at it.
